requirements.txt
discord.py==2.4.0
python-telegram-bot==21.6
python-dotenv==1.0.1
rapidfuzz==3.9.7
